move snapshot to value object
